// frontend/app/src/models/user.ts
export default class User {
    korisnickoIme: string = "";
    lozinka: string= "";
    profilnaSlika : string = "";
    ime: string= "";
    prezime: string= "";
    tip: string= "";
    pol: string= "";
    adresa: string = "";
    kontaktTelefon: string = "";
    email: string = "";
    brojKreditneKartice: string = "";
    odobreno: string = "";
}